wip edlink
