+++ app/core/social-handlers/ClassLinkHandler.js
@@ -1,0 +1,7 @@
+const EdlinkBaseHandler = require('./EdlinkBaseHandler')
+
+module.exports = class ClassLinkHandler extends EdlinkBaseHandler {
+  constructor () {
+    super('classlink')
+  }
+}
