+++ app/styles/modal/auth-modal.sass
@@ -231,6 +231,8 @@
 
         .clever-login-img
           height: 43px
+          background: white
+          padding: 5px
 
       #github-login-button
         position: relative
@@ -248,14 +250,14 @@
       .signup-with-button
         padding: 4px
         display: inline-flex
-        gap: 10px
+        gap: 5px
         justify-content: initial
         align-items: center
         border-radius: 5px
         height: 45px
-        width: 130px        
+        width: 130px
 
-      #schoology-login-btn
+      #schoology-login-btn,#classlink-login-btn
         border: 1px solid #117bbc
         color: black
         background-color: white
