+++ app/styles/modal/create-account-modal/basic-info-view.sass
@@ -34,16 +34,16 @@
       .facebook-logo
         height: 40px
 
-  .schoology-login
-    #schoology-signup-btn
+  .schoology-login,.classlink-login
+    #schoology-signup-btn,#classlink-signup-btn
       padding: 4px
       display: inline-flex
-      gap: 10px
+      gap: 5px
       justify-content: initial
       align-items: center
       border-radius: 5px
       height: 45px
-      width: 100%           
+      width: 100%
       border: 1px solid #117bbc
       color: black
       background-color: white
@@ -58,8 +58,11 @@
 
       .clever-logo
         // Match size of Google button
-        width: 254px
+        width: 154px
         height: 52px
+        background: white
+        padding: 5px
+        border: 1px solid #117bbc
 
   // Forms
