+++ app/templates/core/auth-modal.pug
@@ -78,12 +78,15 @@ mixin auth-form-content
         span Facebook
       if showClever
         a#clever-signup-btn.network-login
-          img.clever-login-img(src="/images/pages/modal/auth/clever_sso_button@2x.png", draggable="false")
+          img.clever-login-img(src="/images/pages/modal/auth/clever-logo-blue.png", draggable="false")
       a#google-login-button.network-login(disabled=true)
         img.google-login-img(src="/images/pages/modal/auth/gplus_sso_button2.svg" draggable="false")
       a#schoology-login-btn.network-login.signup-with-button
-        img.schoology-login-img(src="/images/pages/modal/auth/schoology.png") 
+        img.schoology-login-img(src="/images/pages/modal/auth/schoology.png")
         span Schoology
+      a#classlink-login-btn.network-login.signup-with-button
+        img.classlink-login-img(src="/images/pages/modal/auth/classlink-logo-small.png")
+        span ClassLink
 
   .extra-pane
     if view.utils.isCodeCombat
@@ -116,7 +119,7 @@ mixin auth-form-content
           a#google-login-button(disabled=true, href="#")
             img(src="/images/ozaria/common/log-in-google-sso.svg" draggable="false")
           a#clever-login-btn(href="#")
-            img(src="/images/pages/modal/auth/clever_sso_button@2x.png" draggable="false")    
+            img(src="/images/pages/modal/auth/clever_sso_button@2x.png" draggable="false")
         .row.or-row
           .line
           p.or(data-i18n="login.or")
